requirements.txt
jira==3.5.2
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6